Move `FPoint` and `HPoint` pools next to `LimitingPool`
