+++ pkg/streamingpromql/operator/limiting_pool.go
@@ -6,13 +6,43 @@ import (
 	"github.com/prometheus/prometheus/promql"
 
 	"github.com/grafana/mimir/pkg/util/limiter"
+	"github.com/grafana/mimir/pkg/util/pool"
+)
+
+const (
+	maxExpectedPointsPerSeries  = 100_000 // There's not too much science behind this number: 100000 points allows for a point per minute for just under 70 days.
+	pointsPerSeriesBucketFactor = 2.0
 )
 
 var (
+	fPointSlicePool = pool.NewBucketedPool(1, maxExpectedPointsPerSeries, pointsPerSeriesBucketFactor, func(size int) []promql.FPoint {
+		return make([]promql.FPoint, 0, size)
+	})
+
+	hPointSlicePool = pool.NewBucketedPool(1, maxExpectedPointsPerSeries, seriesPerResultBucketFactor, func(size int) []promql.HPoint {
+		return make([]promql.HPoint, 0, size)
+	})
+
 	// Overrides used only during tests.
 	getFPointSliceForLimitingPool = GetFPointSlice
 	putFPointSliceForLimitingPool = PutFPointSlice
 )
+
+func GetFPointSlice(size int) []promql.FPoint {
+	return fPointSlicePool.Get(size)
+}
+
+func PutFPointSlice(s []promql.FPoint) {
+	fPointSlicePool.Put(s)
+}
+
+func GetHPointSlice(size int) []promql.HPoint {
+	return hPointSlicePool.Get(size)
+}
+
+func PutHPointSlice(s []promql.HPoint) {
+	hPointSlicePool.Put(s)
+}
 
 // LimitingPool manages sample slices for a single query evaluation, and applies any max in-memory samples limit.
 //
