+++ pkg/streamingpromql/operator/pool.go
@@ -9,22 +9,11 @@ import (
 )
 
 const (
-	maxExpectedPointsPerSeries  = 100_000 // There's not too much science behind this number: 100000 points allows for a point per minute for just under 70 days.
-	pointsPerSeriesBucketFactor = 2.0
-
 	maxExpectedSeriesPerResult  = 10_000_000 // Likewise, there's not too much science behind this number: this is the based on examining the largest queries seen at Grafana Labs.
 	seriesPerResultBucketFactor = 2.0
 )
 
 var (
-	fPointSlicePool = pool.NewBucketedPool(1, maxExpectedPointsPerSeries, pointsPerSeriesBucketFactor, func(size int) []promql.FPoint {
-		return make([]promql.FPoint, 0, size)
-	})
-
-	hPointSlicePool = pool.NewBucketedPool(1, maxExpectedPointsPerSeries, seriesPerResultBucketFactor, func(size int) []promql.HPoint {
-		return make([]promql.HPoint, 0, size)
-	})
-
 	matrixPool = pool.NewBucketedPool(1, maxExpectedSeriesPerResult, seriesPerResultBucketFactor, func(size int) promql.Matrix {
 		return make(promql.Matrix, 0, size)
 	})
@@ -47,22 +36,6 @@ var (
 		return nil
 	})
 )
-
-func GetFPointSlice(size int) []promql.FPoint {
-	return fPointSlicePool.Get(size)
-}
-
-func PutFPointSlice(s []promql.FPoint) {
-	fPointSlicePool.Put(s)
-}
-
-func GetHPointSlice(size int) []promql.HPoint {
-	return hPointSlicePool.Get(size)
-}
-
-func PutHPointSlice(s []promql.HPoint) {
-	hPointSlicePool.Put(s)
-}
 
 func GetMatrix(size int) promql.Matrix {
 	return matrixPool.Get(size)
